use shared_ptr

+++ reactor/inetaddress.cpp
@@ -4,10 +4,10 @@ InetAddress::InetAddress() {
 
 }
 
-InetAddress::InetAddress(const std::string &ip, const unsigned short port) {
+InetAddress::InetAddress(const std::string &ip, const uint16_t port) {
     m_addr.sin_family = AF_INET;
     m_addr.sin_addr.s_addr = inet_addr(ip.c_str());
-    m_addr.sin_port = htons(static_cast<uint16_t>(port));
+    m_addr.sin_port = htons(port);
 }
 
 InetAddress::InetAddress(const sockaddr_in &addr): m_addr(addr) {
@@ -22,8 +22,8 @@ const std::string InetAddress::ip() const {
     return std::string(inet_ntoa(m_addr.sin_addr));
 }
 
-const unsigned short InetAddress::port() const {
-    return static_cast<const unsigned short>(ntohs(m_addr.sin_port));
+const uint16_t InetAddress::port() const {
+    return ntohs(m_addr.sin_port);
 }
 
 const sockaddr* InetAddress::addr() const {
